+++ NEWS.md
@@ -22,6 +22,8 @@ incorrectly in previous versions of `jtools`. This has been corrected. Thanks
 to Rebecca Andridge for noticing this. (#89)
 * `wtd.sd()` now gives correct results when data is missing in `x` but not in
 the weights. Thanks to Klaus Langohr for reporting the issue.
+* Users may now choose their own points in `plot_coefs()` and `plot_summs()`
+by passing a vector of shapes to the `point.shape` argument. (#71)
 
 # jtools 2.1.4